rename: @zeltjs/validate-valibot → @zeltjs/validator-valibot
Align with ecosystem conventions (e.g., @hono/valibot-validator)

Verified: precommit-verify ✓ (ed4576d1d60d5606)
