+++ examples/drizzle-todo/package.json
@@ -15,7 +15,7 @@
     "better-sqlite3": "12.9.0",
     "drizzle-orm": "0.45.2",
     "valibot": "1.3.1",
-    "@zeltjs/validate-valibot": "workspace:*",
+    "@zeltjs/validator-valibot": "workspace:*",
     "@zeltjs/openapi": "workspace:*"
   },
   "devDependencies": {
