+++ examples/drizzle-todo/src/todo/todo.controller.ts
@@ -9,7 +9,7 @@ import {
   pathParam,
   response,
 } from '@zeltjs/core';
-import { validated } from '@zeltjs/validate-valibot';
+import { validated } from '@zeltjs/validator-valibot';
 import * as v from 'valibot';
 
 import type { Todo } from '../db/schema';
