+++ examples/hello/package.json
@@ -18,7 +18,7 @@
     "@zeltjs/hono-client": "workspace:*",
     "tsdown": "0.21.10",
     "valibot": "1.3.1",
-    "@zeltjs/validate-valibot": "workspace:*"
+    "@zeltjs/validator-valibot": "workspace:*"
   },
   "devDependencies": {
     "hono": "4.12.16",
