+++ examples/hello/src/entry/hello.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, inject, Post, pathParam, response } from '@zeltjs/core';
-import { validated } from '@zeltjs/validate-valibot';
+import { validated } from '@zeltjs/validator-valibot';
 import * as v from 'valibot';
 
 import { HelloService } from './hello.service';
