+++ examples/workers-url-shortener/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "@zeltjs/core": "workspace:*",
     "valibot": "1.3.1",
-    "@zeltjs/validate-valibot": "workspace:*"
+    "@zeltjs/validator-valibot": "workspace:*"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "4.20260506.1",
