+++ packages/rate-limit/package.json
@@ -33,7 +33,7 @@
     "@types/node": "22.19.17",
     "@zeltjs/core": "workspace:*",
     "@zeltjs/kv": "workspace:*",
-    "@zeltjs/validate-valibot": "workspace:*",
+    "@zeltjs/validator-valibot": "workspace:*",
     "valibot": "1.3.1"
   }
 }
