+++ packages/validator-valibot/vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    name: '@zeltjs/validate-valibot',
+    name: '@zeltjs/validator-valibot',
     include: ['src/**/*.test.ts'],
   },
 });
