+++ tsconfig.json
@@ -12,7 +12,7 @@
     { "path": "packages/redis" },
     { "path": "packages/eventbus" },
     { "path": "packages/rate-limit" },
-    { "path": "packages/validate-valibot" }
+    { "path": "packages/validator-valibot" }
   ],
   "files": []
 }
